[target-spec-miette] move code into another file
Going to add the test library to an optional module.
